Backend/requirements.txt
fastapi==0.120.1
uvicorn==0.38.0
pydantic==2.12.3
python-dotenv==1.0.1
SQLAlchemy==2.0.35
psycopg2-binary==2.9.11

langchain==1.0.2
langchain-core==1.0.1
langchain-google-genai==3.0.0

faiss-cpu==1.8.0.post1
pymupdf
